Move OpenMensa model in own package
